load cover from file
If the embedded cover image is not found, you will see a cover image from the jpg png gif file extensions.
First, a cover is loaded from the file 'cover. [jpg, png, gif]'
When it is missing, the first graphic file is loaded with alphabetical sorting.
All functions are case-dependent.