Guarantee tracks belong to current book before rendering them
This defensive code should fix an issue where the listing might include
disks and chapters from multiple books when users select a book, then
quickly hit the back button and select another book. Since we are
populating the listing on a separate thread, this could lead to a race
condition due where tracks from the first book scheduled to be rendered
would basically collide with the rendering of the current book.