Make track rendering race condition less probable
The hack in the previous commit has made a huge difference in addressing
the race condition and preventing tracks from multiple books being mixed
together in the listing. But I could still see a few odd tracks from
book A in the listing for book B - and that issue seems to have
disappeared while testing on an Ubuntu Virtual Machine.