Fixes an issue where the book detail view could only be openend if the book changed
Opening the book detail view was triggered by setting the property book in the book detail view model.
Because we don't trigger any events anymore when the book doesn't actually change, the book detail view wasn't opened anymore.
This fixes the issue temporarily by separating the view change from the data loading.
When merging the new ui this will be handled more elegantly because there is a central state for the view.