Fix crash on startup
This commit fixes the program crashing because of an uncaught exception.
The bug was triggered by app_settings.last_launched_version being
unset - which would always be the case when launching the program for
the first time!

The bug could also be triggered by manually changing the setting
to an invalid value - e.g. by running:
$ gsettings set com.github.geigi.cozy last-launched-version 'ayy lmao'